1) Add view model files on Product to receive the publisher emitted data. 2) Get call from hiring agent, he want to change request to use SwiftUI, which is not idicated in original task requirement. 3) Next Action: convert UIKit stuff to SwiftUI.

+++ DySearchAPI/Models/Product.swift
@@ -6,13 +6,15 @@
 //  Copyright © 2020 Haibo Family. All rights reserved.
 //
 
-import Foundation
+import SwiftUI
 
 struct Product: Codable {
     let brandName: String?
     let productName: String?
     let productCategory: String?
     let inStock: Bool?
-    let price: Float?
+    let price: CGFloat?
     let id: Int?
+    
+    
 }
